fix: change sum type in withdrawals (int -> float64)

+++ internal/gophermart/models/user_withdraw.go
@@ -10,6 +10,6 @@ type UserWithdraw struct {
 	ID          uuid.UUID `db:"id"`
 	UserID      uuid.UUID `db:"user_id"`
 	OrderID     int64     `db:"order_id"`
-	Sum         int       `db:"sum"`
+	Sum         float64   `db:"sum"`
 	ProcessedAt time.Time `db:"processed_at"`
 }
